+++ opentelemetry-proto/src/proto/tonic/opentelemetry.proto.collector.metrics.v1.rs
@@ -5,12 +5,13 @@ pub struct ExportMetricsServiceRequest {
     /// element. Intermediary nodes (such as OpenTelemetry Collector) that receive
     /// data from multiple origins typically batch the data before forwarding further and
     /// in that case this array will contain multiple elements.
-    #[prost(message, repeated, tag="1")]
-    pub resource_metrics: ::prost::alloc::vec::Vec<super::super::super::metrics::v1::ResourceMetrics>,
+    #[prost(message, repeated, tag = "1")]
+    pub resource_metrics: ::prost::alloc::vec::Vec<
+        super::super::super::metrics::v1::ResourceMetrics,
+    >,
 }
 #[derive(Clone, PartialEq, ::prost::Message)]
-pub struct ExportMetricsServiceResponse {
-}
+pub struct ExportMetricsServiceResponse {}
 /// Generated client implementations.
 pub mod metrics_service_client {
     #![allow(unused_variables, dead_code, missing_docs, clippy::let_unit_value)]
@@ -113,7 +114,7 @@ pub mod metrics_service_client {
 pub mod metrics_service_server {
     #![allow(unused_variables, dead_code, missing_docs, clippy::let_unit_value)]
     use tonic::codegen::*;
-    ///Generated trait containing gRPC methods that should be implemented for use with MetricsServiceServer.
+    /// Generated trait containing gRPC methods that should be implemented for use with MetricsServiceServer.
     #[async_trait]
     pub trait MetricsService: Send + Sync + 'static {
         /// For performance reasons, it is recommended to keep this RPC